+++ api/src/laporan/laporan.module.ts
@@ -1,8 +1,8 @@
 import { Module } from "@nestjs/common";
 import { LaporanController } from "./laporan.controller";
 import { LaporanService } from "./laporan.service";
-import { TambahanController } from "./kegiatan-tambahan.controller";
-import { TambahanService } from "./kegiatan-tambahan.service";
+import { TambahanController } from "./tugas-tambahan.controller";
+import { TambahanService } from "./tugas-tambahan.service";
 import { PrismaService } from "../prisma.service";
 
 @Module({
